+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 * CORE: [CHANGE] Added Endpoint\LoggerTrait to apply generic endpoint operation logging
 * CORE: [FIX] readOnly attributes get stripped out from request
 * CORE: [FIX] binary values in Endpoint\Ldap get base64 encoded
+* CORE: [FEATURE] Possibility to set context data within `map` definition in worklow
 * API: [FIX] uncaught exception: Argument 4 passed to Tubee\Rest\v1\Processes::delete() must implement interface MongoDB\BSON\ObjectIdInterface
 * API: [FIX] uncaught exception: Argument 1 passed to Tubee\Secret\Factory::getOne() must implement interface Tubee\ResourceNamespace\ResourceNamespaceInterface, string given at POST /api/v1/secrets
 * API: [FIX] uncaught exception: Undefined variable: job]  [object] (ErrorException(code: 0): Undefined variable: job at POST /api/v1/jobs